Second Walkway Connection
Connected the second walkway with the rest of the level. However, I ran into a bug where the platforms aren't responsive to the player walking over them. Adding rigidbodies to make the collisions work ended up making all the platforms fall the first time, and even after updating the script the platforms won't fall without the player saying on them for a long while.