Progress
Adding in progress so far on skeleton. I've successfully implemented a patrol for the skeleton, the skeleton noticing the player, the skeleton attacking the player and the player's sword killing the skeleton. The main issues are that the skeleton's death animation is bugged and won't complete and that the skeleton does not damage the player when using its sword.